pyproject.toml
[project]
name = "letswatch"
version = "0.1.5"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "orjson>=3.10.18",
    "pandas>=2.2.3",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
